bg sub + dapi reg + bleed tile variation

+++ coppafish/find_spots/base.py
@@ -177,9 +177,9 @@ def load_spot_info(nbp_file: NotebookPage, nbp_basic: NotebookPage) -> dict:
         spot_info = {'spot_yxz': raw.f.arr_0, 'spot_no': raw.f.arr_1, 'isolated': raw.f.arr_2,
                      'completed': raw.f.arr_3}
     else:
-        spot_info = {'spot_yxz': np.zeros((0, 3), dtype=np.int32),
+        spot_info = {'spot_yxz': np.zeros((0, 3), dtype=np.int16),
                      'spot_no': np.zeros((nbp_basic.n_tiles, nbp_basic.n_rounds + nbp_basic.n_extra_rounds,
-                                          nbp_basic.n_channels), dtype=np.uint16),
+                                          nbp_basic.n_channels), dtype=np.uint32),
                      'isolated': np.zeros((0), dtype=bool),
                      'completed': np.zeros((nbp_basic.n_tiles, nbp_basic.n_rounds + nbp_basic.n_extra_rounds,
                                             nbp_basic.n_channels), dtype=bool)}
